requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.3
cx_Oracle==8.3.0
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-db-logger==0.1.13
django-templated-mail==1.1.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.2.3
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
numpy==2.3.2
oauthlib==3.2.2
openpyxl==3.1.5
oracledb==3.1.1
packaging==25.0
pandas==2.3.1
prompt_toolkit==3.0.52
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
redis==6.4.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.2
xlsxwriter==3.2.5
